Make nginx available in nix flake check test runs
The source tree filtering also now includes the `nginx.conf.template`
file.

If `doCheck = false` is removed from the crane package builds for
ohttp-relay the integration tests are not run by default, so a
nativeBuildInput with nginx is not strictly required even if the
derivation is overridden to reenable the checks.