remove unnecessary #[inline] directives
one of these traces back to the initial commit with no explanations,
most of the others were only added to the persister. even if we had a
good reason to try to optimize this (i.e. a real performance problem),
manual inlining directives should be added using supporting evidence
that the problem is real and the annotation actually achieves the
intended effect, as the compiler is likely to make better decisions.