flake: build dev profile instead of release
this results in much faster builds and therefore checks, improving
developer experience.

if we additionally want to build the packages in release mode, or run
certain tests in release mode, we can build the workspace dependencies
using both release and dev profiles as appropriate, but unless we intend
to ship e.g. docker containers based on our built packages we don't need
the release profile for the foreseeable future.